Authorization for Controllers -byMacMie

+++ src/main/java/com/macmie/mfoodyex/Controller/CreditCardMfoodyController.java
@@ -10,12 +10,16 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.annotation.Secured;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
 import javax.transaction.Transactional;
+import java.security.Principal;
 import java.util.List;
 
+import static com.macmie.mfoodyex.Constant.SecurityConstants.ROLE_ADMIN_SECURITY;
+import static com.macmie.mfoodyex.Constant.SecurityConstants.ROLE_USER_SECURITY;
 import static com.macmie.mfoodyex.Constant.ViewConstants.*;
 
 /*
@@ -35,8 +39,13 @@ public class CreditCardMfoodyController {
     @Autowired
     private UserMfoodyInterfaceService userMfoodyInterfaceService;
 
+    @Autowired
+    private ApplicationCheckAuthorController applicationCheckAuthorController;
+
+    @Secured({ROLE_ADMIN_SECURITY})
     @GetMapping(URL_GET_ALL)
-    public ResponseEntity<?> getAllCreditCardMfoodys() {
+    public ResponseEntity<?> getAllCreditCardMfoodys(Principal principal) {
+        log.info("Get List of CreditCardMfoodys by " + principal.getName());
         List<CreditCardMfoody> creditCardMfoodyList = creditCardMfoodyInterfaceService.getListCreditCardMfoodys();
         if (creditCardMfoodyList.isEmpty()) {
             return new ResponseEntity<>("NO_CONTENT List of CreditCards", HttpStatus.NO_CONTENT);
@@ -44,31 +53,55 @@ public class CreditCardMfoodyController {
         return new ResponseEntity<>(creditCardMfoodyList, HttpStatus.OK);
     }
 
+    @Secured({ROLE_ADMIN_SECURITY, ROLE_USER_SECURITY})
     @GetMapping(URL_GET_BY_ID)
-    public ResponseEntity<?> getCreditCardMfoodyByID(@PathVariable("ID") int ID) {
+    public ResponseEntity<?> getCreditCardMfoodyByID(@PathVariable("ID") int ID, Principal principal) {
+        log.info("Get CreditCardMfoody with ID: {} by {}", ID, principal.getName());
         CreditCardMfoody creditCardMfoody = creditCardMfoodyInterfaceService.getCreditCardMfoodyByID(ID);
         if (creditCardMfoody == null) {
             return new ResponseEntity<>("NOT_FOUND CreditCardMfoody with ID: " + ID, HttpStatus.NOT_FOUND);
         }
+
+        // Check if the current UserMfoody has role ADMIN or the owner of the CreditCardMfoody
+        if(!applicationCheckAuthorController.checkAuthorization(principal, creditCardMfoody.getUser().getIdUser())){
+            return  new ResponseEntity<>("FORBIDDEN Authorization failed!", HttpStatus.FORBIDDEN);
+        }
+
         return new ResponseEntity<>(creditCardMfoody, HttpStatus.OK);
     }
 
+    @Secured({ROLE_ADMIN_SECURITY, ROLE_USER_SECURITY})
     @GetMapping(URL_GET_BY_ID_USER)
-    public ResponseEntity<?> getCreditCardMfoodyByIdUser(@PathVariable("ID") int ID) {
+    public ResponseEntity<?> getAllCreditCardMfoodysByIdUser(@PathVariable("ID") int ID, Principal principal) {
+        log.info("Get List CreditCards with idUser: {} by {}", ID, principal.getName());
         if (userMfoodyInterfaceService.getUserMfoodyByID(ID) == null) {
             return new ResponseEntity<>("NOT_FOUND UserMfoody with idUser: " + ID, HttpStatus.NOT_FOUND);
         }
+
+        // Check if the current UserMfoody has role ADMIN or the owner of the CreditCardMfoody
+        if(!applicationCheckAuthorController.checkAuthorization(principal, ID)){
+            return  new ResponseEntity<>("FORBIDDEN Authorization failed!", HttpStatus.FORBIDDEN);
+        }
+
         List<CreditCardMfoody> creditCardMfoodyList = creditCardMfoodyInterfaceService.getListCreditCardMfoodysByIdUser(ID);
         if (creditCardMfoodyList.isEmpty()) {
-            return new ResponseEntity<>("NO_CONTENT CreditCard of UserMfoody with ID: " + ID, HttpStatus.NO_CONTENT);
+            return new ResponseEntity<>("NO_CONTENT List CreditCards of UserMfoody with idUser: " + ID, HttpStatus.NO_CONTENT);
         }
         return new ResponseEntity<>(creditCardMfoodyList, HttpStatus.OK);
     }
 
+    @Secured({ROLE_ADMIN_SECURITY, ROLE_USER_SECURITY})
     @DeleteMapping(URL_DELETE)
-    public ResponseEntity<?> deleteCreditCardMfoodyByID(@PathVariable("ID") int ID) {
-        if (creditCardMfoodyInterfaceService.getCreditCardMfoodyByID(ID) == null) {
+    public ResponseEntity<?> deleteCreditCardMfoodyByID(@PathVariable("ID") int ID, Principal principal) {
+        log.info("Delete CreditCardMfoody with ID: {} by {}", ID, principal.getName());
+        CreditCardMfoody creditCardMfoody = creditCardMfoodyInterfaceService.getCreditCardMfoodyByID(ID);
+        if (creditCardMfoody == null) {
             return new ResponseEntity<>("NOT_FOUND CreditCardMfoody with ID: " + ID, HttpStatus.NOT_FOUND);
+        }
+
+        // Check if the current UserMfoody has role ADMIN or the owner of the CreditCardMfoody
+        if(!applicationCheckAuthorController.checkAuthorization(principal, creditCardMfoody.getUser().getIdUser())){
+            return  new ResponseEntity<>("FORBIDDEN Authorization failed!", HttpStatus.FORBIDDEN);
         }
 
         try {
@@ -79,30 +112,37 @@ public class CreditCardMfoodyController {
             throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR,
                     "INTERNAL_SERVER_ERROR Exceptions occur when deleting CreditCardMfoody");
         }
-
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
+    @Secured({ROLE_ADMIN_SECURITY, ROLE_USER_SECURITY})
     @DeleteMapping(URL_DELETE_BY_ID_USER)
-    public ResponseEntity<?> deleteCreditCardMfoodyByIdUser(@PathVariable("ID") int ID) {
+    public ResponseEntity<?> deleteAllCreditCardMfoodysByIdUser(@PathVariable("ID") int ID, Principal principal) {
+        log.info("Delete List CreditCardMfoodys with idUser: {} by {}", ID, principal.getName());
         if (userMfoodyInterfaceService.getUserMfoodyByID(ID) == null) {
             return new ResponseEntity<>("NOT_FOUND UserMfoody with ID: " + ID, HttpStatus.NOT_FOUND);
+        }
+
+        // Check if the current UserMfoody has role ADMIN or the owner of the CreditCardMfoody
+        if(!applicationCheckAuthorController.checkAuthorization(principal, ID)){
+            return  new ResponseEntity<>("FORBIDDEN Authorization failed!", HttpStatus.FORBIDDEN);
         }
 
         try {
             creditCardMfoodyInterfaceService.deleteAllCreditCardsMfoodyByIdUser(ID);
         } catch (Exception e) {
-            log.error("An error occurred while deleting CreditCardMfoody with idUser: " + ID);
+            log.error("An error occurred while deleting List CreditCardMfoodys with idUser: " + ID);
             log.error("Detail Error: " + e);
             throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR,
-                    "INTERNAL_SERVER_ERROR Exceptions occur when deleting CreditCardMfoody");
+                    "INTERNAL_SERVER_ERROR Exceptions occur when deleting List of CommentMfoodys");
         }
 
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
-    @PutMapping(URL_EDIT) // idUser is ignored
-    public ResponseEntity<?> editCreditCardMfoody(@RequestBody String creditCardPOJOJsonObject) {
+    @Secured({ROLE_ADMIN_SECURITY, ROLE_USER_SECURITY})
+    @PutMapping(URL_EDIT) // idUser in Json is ignored
+    public ResponseEntity<?> editCreditCardMfoody(@RequestBody String creditCardPOJOJsonObject, Principal principal) {
         try {
             // Convert JsonObject to CreditCardPOJO object, Check input idCard and attach UserMfoody to CreditCardMfoody
             Gson gson = new Gson();
@@ -113,6 +153,12 @@ public class CreditCardMfoodyController {
                 return new ResponseEntity<>("NOT_FOUND UserMfoody with idCard: " + newCreditCardPOJO.getIdCard(),
                         HttpStatus.NOT_FOUND);
             }
+
+            // Check if the current UserMfoody has role ADMIN or the owner of the CreditCardMfoody
+            if(!applicationCheckAuthorController.checkAuthorization(principal, attachUserMfoody.getIdUser())){
+                return  new ResponseEntity<>("FORBIDDEN Authorization failed!", HttpStatus.FORBIDDEN);
+            }
+
             newCreditCardMfoody.setUser(attachUserMfoody);
 
             // Save to DB (Handle Exception in case the unique attributes in the request already exist)
@@ -131,8 +177,9 @@ public class CreditCardMfoodyController {
         }
     }
 
-    @PostMapping(URL_ADD) // idUser must be accurate
-    public ResponseEntity<?> addNewCreditCardMfoody(@RequestBody String creditCardPOJOJsonObject) {
+    @Secured({ROLE_ADMIN_SECURITY, ROLE_USER_SECURITY})
+    @PostMapping(URL_ADD) // idUser in Json must be accurate
+    public ResponseEntity<?> addNewCreditCardMfoody(@RequestBody String creditCardPOJOJsonObject, Principal principal) {
         try {
             // Convert JsonObject to CreditCardPOJO object
             Gson gson = new Gson();
