+++ src/main/java/com/macmie/mfoodyex/Controller/OrderMfoodyController.java
@@ -11,12 +11,16 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.annotation.Secured;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
 import javax.transaction.Transactional;
+import java.security.Principal;
 import java.util.List;
 
+import static com.macmie.mfoodyex.Constant.SecurityConstants.ROLE_ADMIN_SECURITY;
+import static com.macmie.mfoodyex.Constant.SecurityConstants.ROLE_USER_SECURITY;
 import static com.macmie.mfoodyex.Constant.ViewConstants.*;
 
 /*
@@ -39,8 +43,13 @@ public class OrderMfoodyController {
     @Autowired
     private DetailProductOrderMfoodyInterfaceService detailProductOrderMfoodyInterfaceService;
 
+    @Autowired
+    private ApplicationCheckAuthorController applicationCheckAuthorController;
+
+    @Secured({ROLE_ADMIN_SECURITY})
     @GetMapping(URL_GET_ALL)
-    public ResponseEntity<?> getAllOrderMfoodys() {
+    public ResponseEntity<?> getAllOrderMfoodys(Principal principal) {
+        log.info("Get List of OrderMfoodys by " + principal.getName());
         List<OrderMfoody> orderMfoodyList = orderMfoodyInterfaceService.getListOrderMfoodys();
         if (orderMfoodyList.isEmpty()) {
             return new ResponseEntity<>("NO_CONTENT List of OrderMfoodys", HttpStatus.NO_CONTENT);
@@ -48,20 +57,36 @@ public class OrderMfoodyController {
         return new ResponseEntity<>(orderMfoodyList, HttpStatus.OK);
     }
 
+    @Secured({ROLE_ADMIN_SECURITY, ROLE_USER_SECURITY})
     @GetMapping(URL_GET_BY_ID)
-    public ResponseEntity<?> getOrderMfoodyByID(@PathVariable("ID") int ID) {
-        OrderMfoody OrderMfoody = orderMfoodyInterfaceService.getOrderMfoodyByID(ID);
-        if (OrderMfoody == null) {
+    public ResponseEntity<?> getOrderMfoodyByID(@PathVariable("ID") int ID, Principal principal) {
+        log.info("Get OrderMfoody with ID: {} by {}", ID, principal.getName());
+        OrderMfoody orderMfoody = orderMfoodyInterfaceService.getOrderMfoodyByID(ID);
+        if (orderMfoody == null) {
             return new ResponseEntity<>("NOT_FOUND OrderMfoody with ID: " + ID, HttpStatus.NOT_FOUND);
         }
-        return new ResponseEntity<>(OrderMfoody, HttpStatus.OK);
+
+        // Check if the current UserMfoody has role ADMIN or the owner of the OrderMfoody
+        if(!applicationCheckAuthorController.checkAuthorization(principal, orderMfoody.getUser().getIdUser())){
+            return  new ResponseEntity<>("FORBIDDEN Authorization failed!", HttpStatus.FORBIDDEN);
+        }
+
+        return new ResponseEntity<>(orderMfoody, HttpStatus.OK);
     }
 
+    @Secured({ROLE_ADMIN_SECURITY, ROLE_USER_SECURITY})
     @GetMapping(URL_GET_BY_ID_USER)
-    public ResponseEntity<?> getOrderMfoodyByIdUser(@PathVariable("ID") int ID) {
+    public ResponseEntity<?> getAllOrderMfoodysByIdUser(@PathVariable("ID") int ID, Principal principal) {
+        log.info("Get List of OrderMfoodys with idUser: {} by {}", ID, principal.getName());
         if (userMfoodyInterfaceService.getUserMfoodyByID(ID) == null) {
             return new ResponseEntity<>("NOT_FOUND UserMfoody with ID: " + ID, HttpStatus.NOT_FOUND);
         }
+
+        // Check if the current UserMfoody has role ADMIN or the owner of the OrderMfoody
+        if(!applicationCheckAuthorController.checkAuthorization(principal, ID)){
+            return  new ResponseEntity<>("FORBIDDEN Authorization failed!", HttpStatus.FORBIDDEN);
+        }
+
         List<OrderMfoody> orderMfoodyList = orderMfoodyInterfaceService.getListOrderMfoodysByIdUser(ID);
         if (orderMfoodyList.isEmpty()) {
             return new ResponseEntity<>("NO_CONTENT List of OrderMfoodys with idUser: " + ID,
@@ -70,10 +95,18 @@ public class OrderMfoodyController {
         return new ResponseEntity<>(orderMfoodyList, HttpStatus.OK);
     }
 
+    @Secured({ROLE_ADMIN_SECURITY, ROLE_USER_SECURITY})
     @DeleteMapping(URL_DELETE)
-    public ResponseEntity<?> deleteOrderMfoodyByID(@PathVariable("ID") int ID) {
-        if (orderMfoodyInterfaceService.getOrderMfoodyByID(ID) == null) {
+    public ResponseEntity<?> deleteOrderMfoodyByID(@PathVariable("ID") int ID, Principal principal) {
+        log.info("Delete OrderMfoody with ID: {} by {}", ID, principal.getName());
+        OrderMfoody orderMfoody = orderMfoodyInterfaceService.getOrderMfoodyByID(ID);
+        if (orderMfoody == null) {
             return new ResponseEntity<>("NOT_FOUND OrderMfoody with ID: " + ID, HttpStatus.NOT_FOUND);
+        }
+
+        // Check if the current UserMfoody has role ADMIN or the owner of the OrderMfoody
+        if(!applicationCheckAuthorController.checkAuthorization(principal, orderMfoody.getUser().getIdUser())){
+            return  new ResponseEntity<>("FORBIDDEN Authorization failed!", HttpStatus.FORBIDDEN);
         }
 
         try {
@@ -88,19 +121,26 @@ public class OrderMfoodyController {
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
+    @Secured({ROLE_ADMIN_SECURITY, ROLE_USER_SECURITY})
     @DeleteMapping(URL_DELETE_BY_ID_USER)
-    public ResponseEntity<?> deleteOrderMfoodyByIdUser(@PathVariable("ID") int ID) {
+    public ResponseEntity<?> deleteAllOrderMfoodysByIdUser(@PathVariable("ID") int ID, Principal principal) {
+        log.info("Delete List of OrderMfoodys with idUser: {} by {}", ID, principal.getName());
         if (userMfoodyInterfaceService.getUserMfoodyByID(ID) == null) {
             return new ResponseEntity<>("NOT_FOUND UserMfoody with ID: " + ID, HttpStatus.NOT_FOUND);
+        }
+
+        // Check if the current UserMfoody has role ADMIN or the owner of the OrderMfoody
+        if(!applicationCheckAuthorController.checkAuthorization(principal, ID)){
+            return  new ResponseEntity<>("FORBIDDEN Authorization failed!", HttpStatus.FORBIDDEN);
         }
 
         try {
             orderMfoodyInterfaceService.deleteAllOrderMfoodysByIdUser(ID);
         } catch (Exception e) {
-            log.error("An error occurred while deleting OrderMfoody with idUser: " + ID);
+            log.error("An error occurred while deleting List of OrderMfoodys with idUser: " + ID);
             log.error("Detail Error: " + e);
             throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR,
-                    "INTERNAL_SERVER_ERROR Exceptions occur when deleting OrderMfoody");
+                    "INTERNAL_SERVER_ERROR Exceptions occur when deleting List of OrderMfoodys");
         }
 
         return new ResponseEntity<>(HttpStatus.OK);
@@ -113,8 +153,9 @@ public class OrderMfoodyController {
      * 3. The attributes quantityAllProductsInOrder/totalSalPrice/totalFullPrice will get default values 0 and will be
      *    updated later with APIs of DetailProductOrderMfoody coz the oneToMany table is always created first
      * */
+    @Secured({ROLE_ADMIN_SECURITY, ROLE_USER_SECURITY})
     @PutMapping(URL_EDIT)
-    public ResponseEntity<?> editOrderMfoody(@RequestBody String orderPOJOJsonObject) {
+    public ResponseEntity<?> editOrderMfoody(@RequestBody String orderPOJOJsonObject, Principal principal) {
         try {
             // Convert JsonObject to OrderPOJO object, Check input idOrder and set default values for newOrderMfoody
             Gson gson = new Gson();
@@ -123,10 +164,6 @@ public class OrderMfoodyController {
                 return new ResponseEntity<>("NOT_FOUND OrderMfoody with ID: " + newOrderPOJO.getIdOrder(),
                         HttpStatus.NOT_FOUND);
             }
-            OrderMfoody newOrderMfoody = newOrderPOJO.renderOrderMfoody();
-            newOrderMfoody.setQuantityAllProductsInOrder(0);
-            newOrderMfoody.setTotalFullPriceOrder(0);
-            newOrderMfoody.setTotalSalePriceOrder(0);
 
             // Check UserMfoody and set it to OrderMfoody
             UserMfoody attachUserMfoody = userMfoodyInterfaceService.getUserMfoodyByIdOrder(newOrderPOJO.getIdOrder());
@@ -134,6 +171,16 @@ public class OrderMfoodyController {
                 return new ResponseEntity<>("NOT_FOUND UserMfoody with idOrder: " + newOrderPOJO.getIdOrder(),
                         HttpStatus.NOT_FOUND);
             }
+
+            // Check if the current UserMfoody has role ADMIN or the owner of the OrderMfoody
+            if(!applicationCheckAuthorController.checkAuthorization(principal, attachUserMfoody.getIdUser())){
+                return  new ResponseEntity<>("FORBIDDEN Authorization failed!", HttpStatus.FORBIDDEN);
+            }
+
+            OrderMfoody newOrderMfoody = newOrderPOJO.renderOrderMfoody();
+            newOrderMfoody.setQuantityAllProductsInOrder(0);
+            newOrderMfoody.setTotalFullPriceOrder(0);
+            newOrderMfoody.setTotalSalePriceOrder(0);
             newOrderMfoody.setUser(attachUserMfoody);
 
             // Save to DB and return
@@ -152,8 +199,9 @@ public class OrderMfoodyController {
      * 3. Every OrderMfoody is created with quantityAllProductsInOrder/totalSalPrice/totalFullPrice = 0 and will be
      *    updated later with APIs of DetailProductOrderMfoody coz the oneToMany table is always created first
      * */
+    @Secured({ROLE_ADMIN_SECURITY, ROLE_USER_SECURITY})
     @PostMapping(URL_ADD)
-    public ResponseEntity<?> addNewOrderMfoody(@RequestBody String orderPOJOJsonObject) {
+    public ResponseEntity<?> addNewOrderMfoody(@RequestBody String orderPOJOJsonObject, Principal principal) {
         try {
             // Convert JsonObject to OrderPOJO object and Add new UserMfoody to OrderMfoody
             Gson gson = new Gson();
